contributing: add tooling for git send-email
Add a gitconfig target in the Makefile to configure a new clone with
sane defaults:

- set subject prefix
- set correct mailing list address
- enable sendemail.validate
- install sendemail-validate hook

The sendemail-validate hook will make a shallow clone of the current
upstream repo, apply every patch on it and run some checks (a stripped
down version of what is run by the upstream CI).

Add a new check-patches script that verifies that the commit message
actually contains something and that the Signed-off-by trailer from the
patch author is present.

Call check-patches in both the CI and the sendemail-validate hook.

Update CONTRIBUTING.md accordingly.

Signed-off-by: Robin Jarry <robin@jarry.cc>
Acked-by: Tim Culverhouse <tim@timculverhouse.com>