attach: don't glob hidden files
Most of the time it is not wanted to attach hidden files, but by default
globbing does include hidden files.

Add a small check that removes hidden files from the results if they are
not explicitly globbed for or inside a hidden directory.

Implements: https://todo.sr.ht/~rjarry/aerc/83
Signed-off-by: Moritz Poldrack <git@moritz.sh>
Tested-by: Bence Ferdinandy <bence@ferdinandy.com>