filters: install them in $PREFIX/libexec/aerc/filters
The filesystem hierarchy standard describes /usr/share as
"Architecture-independent data". This folder is not intended for
executable scripts and especially not for arch specific binary files
(such as the wrap filter).

Lintian reports an error with aerc 0.14.0:

  arch-dependent-file-in-usr-share [usr/share/aerc/filters/wrap]

Which I had to fix by moving the filter into /usr/libexec.

Install all filters into $PREFIX/libexec/aerc/filters and update the
default SearchDirs to look them up in here as well.

Link: https://refspecs.linuxfoundation.org/FHS_3.0/fhs/ch04s11.html
Link: https://salsa.debian.org/go-team/packages/aerc/-/commit/a0ca00260ffd
Signed-off-by: Robin Jarry <robin@jarry.cc>
Acked-by: Moritz Poldrack <moritz@poldrack.dev>