providers/microsoft: add simpler Office365 instructions
With https://github.com/gaoDean/oauthRefreshToken, there is no need for
mutt_oauth2.py and a personal azure app. The script just takes in your
account, and spits out the refresh token.

This is advantageous because since commit 38d63cb7f788 ("imap,smtp:
cache and cycle XOAUTH2 refresh token"), when xoauth2 is used and
a refresh token is passed, the refresh token will self refresh, whereas
with mutt_oauth2.py, you are generating the actual token, which will
expire and there is generally a lot of moving parts. With
refresh_token.py, it basically serves as a jump start, and everything
else is handled natively within aerc.

Signed-off-by: Dean <gao.dean@hotmail.com>
Acked-by: Robin Jarry <robin@jarry.cc>