pgp: automatically determine signer from sender
This patch automatically determines the signing account from the sender
of the email (i.e. the From email header).

To reflect this in the compose view after changing the From field, the
restriction in updateCrypto() to only run the update if the signing key
is empty has been lifted, so the key always gets updated when calling
updateCrypt().

Additionally, a Signer() method has been added to the Composer to avoid
specifying the same logic twice; once in updateCrypto() and once in
WriteMessage().

If the From header is not populated for any reason the Signer() method
falls back to the from field specified in accounts.conf.

Signed-off-by: witcher <witcher@wiredspace.de>
Acked-by: Robin Jarry <robin@jarry.cc>