templates: avoid error when using .Style with unknown style name
When using {{.Style "foo" bar}} and "foo" is not defined in the styleset
under the [user] section, there is an obscure error:

 template: column-right:1:18: executing "column-right" at <.Style>:
 error calling Style: runtime error: invalid memory address or nil
 pointer dereference

Which is most of the time truncated and can only be read in logs.

Referencing an undefined style is not a serious problem. Return the
default tcell style in that case and avoid the error.

Signed-off-by: Robin Jarry <robin@jarry.cc>
Acked-by: Tim Culverhouse <tim@timculverhouse.com>