templates: add conditional style method
Add .StyleSwitch that takes a string and an arbitrary number of cases
(regexp, style). Reuse some of the constructs introduced by previous
commit.

The style of the first regular expression that matches will be applied.
If the string does not match any of the expressions, it will be left
as-is.

Signed-off-by: Robin Jarry <robin@jarry.cc>
Acked-by: Tim Culverhouse <tim@timculverhouse.com>