ipc: change protocol to JSON
In overhauling the IPC, it has become necessary to switch to a more
extendable message format, to ensure more complex commands can be sent.

Messages have the following basic structure and must not contain
linebreaks, as these are used to delimit separate messages from one
another.

    {"arguments": ["mailto:moritz@poldrack.dev"]}

The responses have the following structure:

    {"error": "epic fail"}

If the IPC request was successful, "error" will be empty.

    {"error": ""}

Signed-off-by: Moritz Poldrack <git@moritz.sh>
Signed-off-by: Robin Jarry <robin@jarry.cc>