contrib: fix sendemail validate script
The current sendemail-validate script cannot be used for patch series.
It works on each patch independently from the others and make it
impossible to use when there are inter dependencies.

I have submitted an alternate validate script that works on whole series
which is still waiting for reviews. In the meantime, propose to use my
script which can work. To install it in your environment:

 curl -Lo ~/.local/bin/git-send-email \
   https://paste.sr.ht/blob/cedcf24383022949c8dc5bc1300cf5cbc879b36f
 echo 'export GIT_EXEC_PATH=$PATH:$(git --exec-path)' >> ~/.profile

Link: https://lore.kernel.org/git/20230103231133.64050-1-robin@jarry.cc/
Link: https://paste.sr.ht/~rjarry/7e9a98866ba2c6f34e4169debf4f2c14b574613e
Signed-off-by: Robin Jarry <robin@jarry.cc>