term: fix handling of failed command
A panic occurs when a terminal is launched with a command that fails:
	:term blabla

The underlying terminal does not need to be closed if the command failed
to start: all resources are cleaned up upon failure to start already.
Don't attempt to close terminal if the command didn't start.

With the above fix in place, the tab will linger around until a redraw
occurs, as there is nothing queuing a redraw on this behavior. Add a
QueueRedraw in the tabs.Remove method to clean up the tabbar.

The issue in tcell-term that causes this panic has been address there as
well, released as 0.7.1. Update aerc's version.

Reported-by: Koni Marti <koni.marti@gmail.com>
Signed-off-by: Tim Culverhouse <tim@timculverhouse.com>
Tested-by: Koni Marti <koni.marti@gmail.com>
Acked-by: Robin Jarry <robin@jarry.cc>