filters: fix option parsing on arm cpus
When running colorize or wrap on ARM, both programs exit immediately
displaying their help message whatever the provided arguments. This is
caused by an implicit downcast of the getopt return value.

On most architectures, char is signed by default. On ARM, char is
unsigned for performance reasons. Since the signed int return value of
getopt is forced into a char, the results differ on ARM compared to x86.

* Add -Wconversion -Warith-conversion to CFLAGS in CI builds to ensure
  catching such issues in the future.

* Fix all -Wconversion -Warith-conversion reported issues.

* Wide char functions need to deal with wint_t and wchar_t and it is
  guaranteed that a wchar_t can always fit into a wint_t. Add explicit
  casts to silence the reported warnings.

Link: https://www.arm.linux.org.uk/docs/faqs/signedchar.php
Link: https://lwn.net/Articles/911914/
Fixes: https://todo.sr.ht/~rjarry/aerc/164
Reported-by: Bence Ferdinandy <bence@ferdinandy.com>
Suggested-by: Allen Sobot <chilledfrogs@disroot.org>
Signed-off-by: Robin Jarry <robin@jarry.cc>
Tested-by: Bence Ferdinandy <bence@ferdinandy.com>