dirtree: modify drawing of the dirtree
Update the dirtree drawing structure to be a bit more compact. The
dirtree is drawn with one column left for a flag, which indicates if a
parent is collapsed or expanded. Only draw the "flag" when the parent is
collapsed. Change the flag to a '+' character to match common UI
patterns of expanding lists having a '+' box.

Don't draw the '>' character in the dirtree. This character is nice in
the threaded message view, but clutters up the dirtree. I know this is a
matter of preference, but this approach is similar to most other UI tree
view and gives the dirtree an extra column of space.

The original dirtree looks like this:
	┌Inbox
	├─>Sub
	└─>Sub

And collapsed:
	-Inbox

This patch changes it to:
	Inbox
	├─Sub
	└─Sub

And collapsed:
	+Inbox

Signed-off-by: Tim Culverhouse <tim@timculverhouse.com>
Acked-by: Moritz Poldrack <moritz@poldrack.dev>