save: new option for saving attachments
Add a new `-A` option to `:save` that works in the same manner as `-a`,
but saves all the named parts of an email, not just attachments.

The reason is that I have an email with this structure:

multipart/related
multipart/alternative
      text/plain
      text/html
    image/png (image001.png)
    image/png (image002.png)
    image/png (image003.png)
  text/plain (env.txt)

Where the `env.txt` is a "real" attachment, while the images are just a
part of the HTML version of the email. However, in this particular email
it was important to see them which can't be done with text UI and
opening the HTML part with the browser also didn't work. Saving them to
a temorary folder did the job and this can be useful in other scenarios.

So before the patch we could do `:save -ap /some/path` and get just the
`env.txt` saved there.

After the patch we could also do `:save -Ap /some/path` and get all the
images and the text file saved into the folder.

Signed-off-by: Vitaly Ovchinnikov <v@postbox.nz>
Tested-by: Moritz Poldrack <moritz@poldrack.dev>