logging: remove ultra verbose trace logs
Every message exchanged between the UI and the worker(s) is logged at
least twice with some obscure IDs and struct pointer names.

When enabling trace logging, this floods the files for very little or no
benefit and makes the meaningful logging message hard to catch.

Also, other messages tend to flood the output for no specific reason.
These are most of the time leftovers from the development process and
are not useful for tracking potential issues.

Remove these.

Signed-off-by: Robin Jarry <robin@jarry.cc>
Tested-by: Inwit <inwit@sindominio.net>