compose: factorize body read in a method
In preparation for allowing headers in the text editor, put code related
to reading the email body in a method and use it where needed.

Never close the email temp file, always keep it opened in read-write
mode. This will be also required to update embedded header values when
running :edit after modifying email headers from the compose::review
screen.

Signed-off-by: Robin Jarry <robin@jarry.cc>
Tested-by: Bence Ferdinandy <bence@ferdinandy.com>
Tested-by: Koni Marti <koni.marti@gmail.com>